+++ src/main/java/healthcalc/MetabolicMetrics.java
@@ -1,5 +1,5 @@
 package healthcalc;
 
 public interface MetabolicMetrics {
-    public float basalMetabolicRate(float weight, int height, char gender, int age) throws Exception;
+    public double basalMetabolicRate(float weight, int height, char gender, int age) throws Exception;
 }
